+++ src/Commands/ImportAttribute.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace CloseAllTabs
+{
+    internal class ImportAttribute : Attribute
+    {
+    }
+}