+++ src/source.extension.cs
@@ -11,7 +11,7 @@ namespace CloseAllTabs
         public const string Name = "Clean Solution";
         public const string Description = @"Cleans up the solution for the next time you open it. Closes open documents and collapses folders automatically when a solution is being closed. ";
         public const string Language = "en-US";
-        public const string Version = "1.3";
+        public const string Version = "1.4";
         public const string Author = "Mads Kristensen";
         public const string Tags = "tabs, cleanup, collapse";
     }