+++ CHANGELOG.md
@@ -11,6 +11,10 @@ download in the
 These are the changes to each version that has been released
 on the official Visual Studio extension gallery.
 
+## 1.4
+
+- [x] Focus Solution Explorer on close
+
 ## 1.3
 
 - [x] Option to collapse projects